% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_defaults.R
\name{add_defaults_prm.daily}
\alias{add_defaults_prm.daily}
\title{Add defaults for \code{prm.daily}}
\usage{
add_defaults_prm.daily(x)
}
\arguments{
\item{x}{List. \code{prm.daily}.}
}
\value{
List. Updated version of \code{prm.daily}.
}
\description{
Add defaults for \code{prm.daily}
}
\seealso{
estimate_R_ww estimate_R_cl
}
