% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_defaults.R
\name{add_defaults_prm.R}
\alias{add_defaults_prm.R}
\title{Add defaults for \code{prm.R}}
\usage{
add_defaults_prm.R(x)
}
\arguments{
\item{x}{List. \code{prm.R}.}
}
\value{
List. Updated version of \code{prm.R}.
}
\description{
Add defaults for \code{prm.R}
}
\seealso{
estimate_R_ww estimate_R_cl
}
