% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{add_one}
\alias{add_one}
\title{Dummy test function}
\usage{
add_one(x)
}
\arguments{
\item{x}{A numerical value}
}
\value{
a numerical value incremented by 1.
}
\description{
Just a dummy test, this will be erased!
}
\examples{
add_one(3.14)

}
