% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_to_daily.R
\name{attach_startdate_agg}
\alias{attach_startdate_agg}
\title{Attach start date from first observation for aggregated data
from time (day number)}
\usage{
attach_startdate_agg(x)
}
\description{
counterpart function to \code{attach_t_agg()}
}
