% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_t.R
\name{attach_t}
\alias{attach_t}
\title{Attach internal time index column}
\usage{
attach_t(df)
}
\arguments{
\item{df}{Data frame. Must have column \code{date}.}
}
\value{
Tibble.
}
\description{
Attach internal time index column
}
