% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggreg_inc.R
\name{build_A}
\alias{build_A}
\title{Build aggregation matrix A}
\usage{
build_A(N, obs_times, window)
}
\arguments{
\item{N}{Integer. Total number of days after disaggregation.}

\item{obs_times}{Integer vector. Observation times of the aggregated (reported) incidence.}

\item{window}{Integer. Aggregation window size. For example, \code{window = 7} for weekly aggregation.}
}
\description{
Build aggregation matrix A
}
\keyword{internal}
