% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggreg_inc.R
\name{build_D2}
\alias{build_D2}
\title{Build second-difference matrix D, that is:
D2(y) = y(t+1) - 2 y(t) + y(t-1)}
\usage{
build_D2(N)
}
\arguments{
\item{N}{Integer. Total number of days after disaggregation.}
}
\description{
Build second-difference matrix D, that is:
D2(y) = y(t+1) - 2 y(t) + y(t-1)
}
\keyword{internal}
