% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_cl.input_daily}
\alias{check_cl.input_daily}
\title{Check if input clinical data is already daily}
\usage{
check_cl.input_daily(cl.input, silent = FALSE)
}
\arguments{
\item{cl.input}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of report
\item \code{count}: count of reported cases
}}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\value{
Logical. Indicates whether clinical data is already daily.
}
\description{
Check if input clinical data is already daily
}
