% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_cl.input_format}
\alias{check_cl.input_format}
\title{Check the format of input clinical data}
\usage{
check_cl.input_format(cl.data, silent = FALSE)
}
\arguments{
\item{cl.data}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of report
\item \code{value}: count of reported cases
}}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\description{
Check the format of input clinical data
}
\keyword{internal}
