% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_data.cl_format}
\alias{check_data.cl_format}
\title{Check the format of input clinical data}
\usage{
check_data.cl_format(data.cl, silent = FALSE)
}
\arguments{
\item{data.cl}{Data frame. Timeseries of clinical reports. Must have at least \code{date} and \code{count} columns.}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\description{
Check the format of input clinical data
}
