% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_data_clin}
\alias{check_data_clin}
\title{Check the dataframe of clinical data}
\usage{
check_data_clin(cl.agg, silent = FALSE)
}
\arguments{
\item{cl.agg}{Data frame. Must have variables:
\itemize{
\item{\code{date}: }{calendar date of report}
\item{\code{count}: }{count of reported cases}
}}

\item{silent}{Logical. Should functions be run silently, i.e., without any messages and warnings to console?}
}
\description{
Check the dataframe of clinical data
}
