% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_df.input_daily}
\alias{check_df.input_daily}
\title{Check if input data is already daily}
\usage{
check_df.input_daily(df.input)
}
\arguments{
\item{df.input}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of observation (report for clinical data, wastewater sampling for wastewater data)
\item \code{value}: observed values (count of reported cases for clinical data, wastewater concentration for wastewater data)
}}
}
\value{
Logical. Indicates whether input data is already daily.
}
\description{
Check if input data is already daily
}
\keyword{internal}
