% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_dist}
\alias{check_dist}
\title{Check distributions}
\usage{
check_dist(x)
}
\arguments{
\item{x}{family of distributions to be checked, as generated by \code{def_dist_()}}
}
\description{
Check distributions
}
\keyword{internal}
