% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_for_deconv}
\alias{check_for_deconv}
\title{Check that deconvolution inputs are compatible}
\usage{
check_for_deconv(obs, dist)
}
\arguments{
\item{obs}{numeric vector. signal (\emph{e.g.}, case reports)}

\item{dist}{numeric vector. deconvolution kernel (\emph{e.g.}, reporting delay distribution)}
}
\description{
Check that deconvolution inputs are compatible
}
