% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.daily.check}
\alias{check_prm.daily.check}
\title{Check parameters for daily data inference check}
\usage{
check_prm.daily.check(x)
}
\arguments{
\item{x}{List. Parameters for daily data inference check.}
}
\description{
Check parameters for daily data inference check
}
