% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.daily}
\alias{check_prm.daily}
\title{Check parameters for daily data inference}
\usage{
check_prm.daily(x)
}
\arguments{
\item{x}{List. Parameters for daily data inference.}
}
\description{
Check parameters for daily data inference
}
