% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.R}
\alias{check_prm.R}
\title{Check parameters for Rt calculation}
\usage{
check_prm.R(prm.R, silent = FALSE)
}
\arguments{
\item{prm.R}{List. Settings for the ensemble when calculating Rt. Elements include:
\itemize{
\item{\code{iter}: }{number of iterations for ensemble}
\item{\code{window}: }{length of time window to use for each Rt estimate. if \code{t_end} is specified in \code{config.EpiEstim}, this option will override it.}
\item{\code{config.EpiEstim}: }{configuration for \code{EpiEstim} defined via \code{\link[EpiEstim:make_config]{EpiEstim::make_config()}}. if \code{NULL}, will use default config from \code{EpiEstim}.}
}}

\item{silent}{Logical. Should functions be run silently, i.e., without any messages and warnings to console?}
}
\description{
Check parameters for Rt calculation
}
