% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.smooth}
\alias{check_prm.smooth}
\title{Check parameters for smoothing}
\usage{
check_prm.smooth(x)
}
\arguments{
\item{x}{List that specifies the type of smoothing and the parameters associated with the smoothing method.}
}
\description{
Check parameters for smoothing
}
