% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cl.agg}
\alias{cl.agg}
\title{Sample of aggregated clinical reports}
\format{
\subsection{\code{cl.agg}}{

A data frame with 84 rows and 4 columns:
\itemize{
\item \code{pt}: standard two-character abbreviation (lowercase) of the province name (based on \href{https://www12.statcan.gc.ca/census-recensement/2021/ref/dict/tab/index-eng.cfm?ID=T1_8}{Statcan 2021 census abbreviations})
\item \code{date}: report date
\item \code{count}: count of reported cases for the previous week
}
}
}
\usage{
cl.agg
}
\description{
A subset of COVID-19 weekly reports in the Government of Canada Health Infobase
}
\keyword{datasets}
