% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cl.data}
\alias{cl.data}
\title{Sample of aggregated clinical reports}
\format{
\subsection{\code{cl.data}}{

A data frame with 96 rows and 3 columns:
\itemize{
\item \code{pt}: standard two-character abbreviation (lowercase) of the province name (based on Statcan 2021 census abbreviations)
\item \code{date}: report date
\item \code{value}: count of reported cases for the previous week
}

Filter indicating a specific province to extract a sample dataset for use with \code{\link[=estimate_R_cl]{estimate_R_cl()}}, \emph{e.g.}

\code{estimate_R_cl(cl.data = dplyr::filter(cl.data, pt == 'bc'), ...)}
}
}
\usage{
cl.data
}
\description{
A subset of COVID-19 weekly reports in the Government of Canada Health Infobase
}
\keyword{datasets}
