% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{deconv}
\alias{deconv}
\title{Wrapper for deconvolution with a given distribution}
\usage{
deconv(counts, dist, max.iter = 10, silent = FALSE)
}
\arguments{
\item{counts}{Numeric. Vector of daily counts.}

\item{dist}{Numeric. Vector of discrete distributions (daily, truncated) with which we're deconvoluting counts, \emph{e.g.}, produced by \code{\link[=get_discrete_dist]{get_discrete_dist()}}}

\item{max.iter}{Numeric. Maximum number of Richardson-Lucy iterations.}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\description{
Wrapper for deconvolution with a given distribution
}
\keyword{internal}
