% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_ww_inc.R
\name{deconv_ww_inc}
\alias{deconv_ww_inc}
\title{Deconvoluting Wastewater Data to Incidence}
\usage{
deconv_ww_inc(d, fec, scaling.factor)
}
\arguments{
\item{d}{ww dataframe}

\item{fec}{fecal shedding distribution}

\item{scaling.factor}{scaling factor of ww data.}
}
\value{
Dataframe with deconvoluted incidence
}
\description{
Function estimates incidence from smoothed data
}
