% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_example.R
\name{dist_bulk_example}
\alias{dist_bulk_example}
\title{Example of distributions for a given virus}
\usage{
dist_bulk_example(virus)
}
\arguments{
\item{virus}{String. Name of the virus.
Currently implemented values are
\code{COVID}, \code{FLUA}, \code{FLUB}, \code{RSV}}
}
\value{
A list of distributions (generation interval, incubation period,
fecal shedding).
}
\description{
Example of distributions for a given virus
}
\examples{
dist_bulk_example('COVID')


}
