% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrib_example.R
\name{dist_bulk_example_reporting}
\alias{dist_bulk_example_reporting}
\title{Examples for reporting distributions.}
\usage{
dist_bulk_example_reporting(jurisdiction)
}
\arguments{
\item{jurisdiction}{String. Jurisdiction for which the reporting distributions are typical.
Only "Canada" is currently implemented.}
}
\value{
List of reporting distributions (delay and fraction).
}
\description{
Examples for reporting distributions.
}
\examples{
dist_bulk_example_reporting("Canada")

}
