% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw_from_gamma}
\alias{draw_from_gamma}
\title{Draw from gamma based on parameter name in distribution list.}
\usage{
draw_from_gamma(par, dist)
}
\arguments{
\item{par}{String. Name of the parameter to sample.}

\item{dist}{List. Distribution definition.}
}
\value{
Numeric. The sampled value.
}
\description{
Draw from gamma based on parameter name in distribution list.
}
