% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl.R
\name{estimate_R_cl}
\alias{estimate_R_cl}
\title{Estimate the effective reproduction from clinical report data}
\usage{
estimate_R_cl(
  cl.agg,
  dist.repdelay,
  dist.repfrac,
  dist.incub,
  dist.gi,
  popsize,
  prm.daily = list(burn = 500, iter = 2000, chains = 20, first.agg.period = NULL),
  prm.daily.check = list(agg.reldiff.tol = 10),
  prm.smooth = list(window = 7),
  prm.R = list(iter = 10, window = 7, config.EpiEstim = NULL),
  silent = FALSE
)
}
\arguments{
\item{cl.agg}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of report
\item \code{count}: count of reported cases
}}

\item{dist.repdelay}{List. Parameters for the reporting delay distribution in the same format as returned by \code{\link[=def_dist_reporting_delay]{def_dist_reporting_delay()}}.}

\item{dist.repfrac}{List. Parameters for the reporting fraction distribution in the same format as returned by \code{\link[=def_dist_reporting_fraction]{def_dist_reporting_fraction()}}.}

\item{dist.incub}{List. Parameters for the incubation period distribution in the same format as returned by \code{\link[=def_dist_incubation_period]{def_dist_incubation_period()}}.}

\item{dist.gi}{List. Parameters for the generation interval distribution in the same format as returned by \code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}.}

\item{popsize}{Integer. Population size to use in MCMC simulation to infer daily observations from aggregated input data.}

\item{prm.daily}{List. Parameters for daily report inference via MCMC. Elements include:
\itemize{
\item \code{burn}: length of burn-in period (number of days)
\item \code{iter}: number of iterations after burn-in period (number of days)
\item \code{chains}: number of chains to simulate
\item \code{first.agg.period}: length of aggregation period for first aggregated observation (number of days); if NULL, assume same aggregation period as observed for second observation (gap between first and second observations)
}}

\item{prm.daily.check}{List. Parameters for checking aggregated to daily report inference. Elements include:
\itemize{
\item \code{agg.reldiff.tol}: numerical tolerance (\%) for relative error between aggregated inferred daily reports and original aggregated reports; chronological observations are dropped until this tolerance is first acheived (convergence at the start of the timeseries is often the worst, need to maintain uninterrupted daily timeseries for input into Rt calculation).
}
Set this entire argument to \code{NULL} to use inferred daily reports as is.}

\item{prm.smooth}{List. Smoothing parameters for input signal (daily reports) into Rt calculation. Elements include:
\itemize{
\item \code{method}: smoothing method to use; currently only \code{rollmean} (centred rolling average) is implemented
\item \code{window}: width of smoothing window (number of days)
}
Set this entire argument to \code{NULL} to turn off smoothing.}

\item{prm.R}{List. Settings for the ensemble when calculating Rt.
Elements include:
\itemize{
\item \code{window}: number of days defining the window of data used by
\code{EpiEstim} to estimate Rt
\item \code{CI}: Numeric between 0 and 1. Confidence interval width for Rt
estimates after sampling uncertain distributions.
\item \code{config.EpiEstim}: (optional) configuration for \code{EpiEstim} defined via
\code{\link[EpiEstim:make_config]{EpiEstim::make_config()}}. If \code{NULL}, will use default config from
\code{EpiEstim}.
}}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\value{
List. Elements include:
\itemize{
\item \code{cl.agg}: original aggregated reports signal
\item \code{cl.daily}: reports as input for Rt calculation (inferred daily counts, smoothed)
\item \code{R}: the effective R estimate (summary from ensemble)
}
}
\description{
Estimate the effective reproduction from clinical report data
}
