% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{estimate_R_cl_single}
\alias{estimate_R_cl_single}
\title{A single realization of the Rt estimate}
\usage{
estimate_R_cl_single(
  cl.daily,
  dist.repfrac,
  dist.repdelay,
  dist.incub,
  dist.gi
)
}
\arguments{
\item{cl.daily}{dataframe. realizations of daily report counts. must include at least \code{date} and \code{value} columns.}

\item{dist.repfrac}{parameters for the reporting fraction distribution}

\item{dist.repdelay}{parameters for the reporting delay distribution}

\item{dist.incub}{parameters for the incubation period distribution}

\item{dist.gi}{parameters for the generation interval distribution}
}
\description{
A single realization of the Rt estimate
}
