% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcdiag.R
\name{extract_mcmc_values}
\alias{extract_mcmc_values}
\title{Extract MCMC chains from a JAGS object}
\usage{
extract_mcmc_values(chain, jags.obj)
}
\arguments{
\item{chain}{Integer. Chain number.}

\item{jags.obj}{JAGS object as returned by \code{code.sample()}}
}
\value{
A dataframe of the chain values for selected parameters.
}
\description{
Extract MCMC chains from a JAGS object
}
