% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_to_daily.R
\name{fit_jags_aggreg}
\alias{fit_jags_aggreg}
\title{Fit JAGS model to aggregated data}
\usage{
fit_jags_aggreg(
  obs.times,
  Y,
  g,
  N,
  n.days = NULL,
  mcmc.params = list(burn = 1000, iter = 3000, chains = 3),
  path.jags = "R/"
)
}
\description{
Fit JAGS model to aggregated data
}
