% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cl-daily.R
\name{format_cl.daily}
\alias{format_cl.daily}
\title{Format clinical data for estimate_R_cl_rep()}
\usage{
format_cl.daily(cl.input)
}
\arguments{
\item{cl.input}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of report
\item \code{count}: count of reported cases
}}
}
\description{
Format clinical data for estimate_R_cl_rep()
}
