% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecal_shedding.R
\name{get_fecal_shedding}
\alias{get_fecal_shedding}
\title{Fecal Shedding Distribution}
\usage{
get_fecal_shedding(pathogen, subtype = "")
}
\arguments{
\item{pathogen}{String. Name of the pathogen ('sarscov2', 'influenza', 'RSV')}

\item{subtype}{String.}
}
\value{
A numerical vector where each element represent one day since infection.
}
\description{
Fecal Shedding Distribution
}
\examples{
fec = get_fecal_shedding('sarscov2')
print(fec)

}
