% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecal_shedding.R
\name{get_gamma_fec}
\alias{get_gamma_fec}
\title{Fecal Shedding Gamma Distribution}
\usage{
get_gamma_fec(fec.list)
}
\arguments{
\item{fec.list}{List of fecal shedding parameters based on various shedding
profiles (e.g. asymptomatic, symptomatic, and hospitalized symptomatic).}
}
\value{
A dataframe containing the mean and sd of gamma distributions for
each shedding profile.
}
\description{
Function generates a gamma distribution for each fecal shedding
profile.
}
