% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gi.R
\name{get_gi}
\alias{get_gi}
\title{Retrieve the Generation Interval Distribution}
\usage{
get_gi(pathogen, subtype)
}
\arguments{
\item{pathogen}{String. Name of the pathogen.}

\item{subtype}{String. Oprion subtype}
}
\value{
A list.
}
\description{
Retrieve the Generation Interval Distribution
}
\examples{
gi = get_gi('sarscov2')
print(gi)

}
