% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_to_daily.R
\name{get_realizations}
\alias{get_realizations}
\title{Retrieve realizations for weekly -> daily inference}
\usage{
get_realizations(fit.reports.daily, reports)
}
\arguments{
\item{fit.reports.daily}{dataframe. realizations from daily report inference. must at least have \code{t} (time index), \code{var} (variable name), \code{iteration} (realization number), and \code{value} (inferred count) columns.}

\item{reports}{dataframe. original weekly reports. must at least have \code{date} column}
}
\description{
Retrieve realizations for weekly -> daily inference
}
\seealso{
\code{\link[=weekly_to_daily]{weekly_to_daily()}}
}
