% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_tnorm_bounds}
\alias{get_tnorm_bounds}
\title{Get non-negative truncated normal bounds}
\usage{
get_tnorm_bounds(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{sd}{standard deviation}
}
\description{
Get non-negative truncated normal bounds
}
