% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_ww.R
\name{inc2R_one_gi}
\alias{inc2R_one_gi}
\title{Helper function.
Converts incidence to Rt after sampling one generation interval distribution}
\usage{
inc2R_one_gi(i, dist.gi, incidence, prm.R)
}
\arguments{
\item{i}{Iteration index. (not used but required when using \code{lapply()})}

\item{dist.gi}{Generation interval distribution.}

\item{incidence}{Incidence object.}

\item{prm.R}{List of configuration parameters for EpiEstim.}
}
\description{
Helper function.
Converts incidence to Rt after sampling one generation interval distribution
}
