% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_ww.R
\name{inc2R_one_iter}
\alias{inc2R_one_iter}
\title{Helper function.
Converts wastewater to Rt after sampling one fecal shedding and
one generation interval distribution.}
\usage{
inc2R_one_iter(i, dist.fec, dist.gi, wastewater, scaling.factor, prm.R)
}
\arguments{
\item{i}{Iteration index. (not used but required when using \code{lapply()})}

\item{dist.fec}{Fecal shedding distribution.}

\item{dist.gi}{Generation interval distribution.}

\item{wastewater}{Smoothed wastewater time-series.}

\item{scaling.factor}{Scaling from wastewater concentration to prevalence.
This value may be assumed or independently calibrated to data.}

\item{prm.R}{List of configuration parameters for EpiEstim.}
}
\description{
Helper function.
Converts wastewater to Rt after sampling one fecal shedding and
one generation interval distribution.
}
