% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_to_R.R
\name{inc_to_R}
\alias{inc_to_R}
\title{Estimating Rt from incidence}
\usage{
inc_to_R(df, start.date, gi)
}
\arguments{
\item{df}{dataframe}

\item{start.date}{start date of data}

\item{gi}{generation interval}
}
\value{
dataframe with Rt estimates
}
\description{
Function estimates Rt from incidence data
}
