% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_int_daily.R
\name{linear_int_daily}
\alias{linear_int_daily}
\title{Daily incidence from linear interpolation}
\usage{
linear_int_daily(cl.input)
}
\arguments{
\item{cl.input}{Aggregated incidence.}
}
\value{
A dataframe of daily incidence
}
\description{
Daily incidence from linear interpolation
}
