% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diagnostic_ww}
\alias{plot_diagnostic_ww}
\title{Diagnostic plot for R estimation from wastewater data}
\usage{
plot_diagnostic_ww(r.estim, caption = NULL)
}
\arguments{
\item{r.estim}{List. Output of \code{\link[=estimate_R_ww]{estimate_R_ww()}}.}

\item{caption}{Character. Optional plot caption.}
}
\value{
A \code{ggplot} object.
}
\description{
Diagnostic plot for R estimation from wastewater data
}
\examples{

# Load data of viral concentration in wastewater
data("ww.data")

# Estimate Rt based on wastewater data
x = estimate_R_ww(
  ww.conc  = ww.data,
  dist.fec = ern::def_dist(
    dist = "gamma",
    mean = 12.90215,
    mean_sd = 1.136829,
    shape = 1.759937,
    shape_sd = 0.2665988,
    max = 33
    ),
  dist.gi  = ern::def_dist(
    dist     = "gamma",
    mean     = 6.84,
    mean_sd  = 0.7486,
    shape    = 2.39,
    shape_sd = 0.3573,
    max      = 15
    ), 
  silent   = TRUE
)

# Diagnostic plot
g = plot_diagnostic_ww(x)
plot(g)


}
\seealso{
\code{\link[=estimate_R_ww]{estimate_R_ww()}} \code{\link[=plot_diagnostic_cl]{plot_diagnostic_cl()}}
}
