% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcdiag.R
\name{plot_gelman_rubin}
\alias{plot_gelman_rubin}
\title{Plot the Gelman Rubin statistic for all parameters.}
\usage{
plot_gelman_rubin(jags.obj)
}
\arguments{
\item{jags.obj}{JAGS object as returned by \code{code.sample()}}
}
\value{
A \code{ggplot} plot.
}
\description{
Plot the Gelman Rubin statistic for all parameters.
}
