% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcdiag.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot MCMC traces}
\usage{
plot_traces(jags.obj)
}
\arguments{
\item{jags.obj}{JAGS object as returned by \code{code.sample()}}
}
\value{
A \code{ggplot} plot.
}
\description{
Plot MCMC traces
}
