% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{reports_to_incidence}
\alias{reports_to_incidence}
\title{Infer incidence from reports via deconvolutions}
\usage{
reports_to_incidence(
  reports.daily,
  reporting.delay,
  incubation.period,
  max.iter = 50
)
}
\description{
Infer incidence from reports via deconvolutions
}
