% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{reshape_fit_jags}
\alias{reshape_fit_jags}
\title{Reshape JAGS fit object}
\usage{
reshape_fit_jags(x)
}
\arguments{
\item{x}{dataframe. JAGS output from \code{\link[=fit_jags_aggreg]{fit_jags_aggreg()}}.}
}
\description{
Reshape JAGS fit object
}
\seealso{
\code{\link[=fit_jags_aggreg]{fit_jags_aggreg()}}
}
