% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sample_a_dist}
\alias{sample_a_dist}
\title{Sample parameters for a single distribution from
a family of distributions, assuming parameters come from a Gamma distribution}
\usage{
sample_a_dist(dist)
}
\arguments{
\item{dist}{List. A list of distribution parameters, as defined by the
\verb{def_dist_*()} functions.}
}
\description{
Sample parameters for a single distribution from
a family of distributions, assuming parameters come from a Gamma distribution
}
