% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{sample_from_dist}
\alias{sample_from_dist}
\title{Sample from a distribution
(currently only implemented for a uniform distribution)}
\usage{
sample_from_dist(n, params)
}
\arguments{
\item{n}{number of samples to draw}

\item{params}{distribution parameters}
}
\description{
Sample from a distribution
(currently only implemented for a uniform distribution)
}
\keyword{internal}
