% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_to_R.R
\name{sample_post_ee}
\alias{sample_post_ee}
\title{Helper function - sample from EpiEstim's R posterior distribution}
\usage{
sample_post_ee(window, objee, n)
}
\arguments{
\item{window}{integer. Time slice used to estimate Rt.}

\item{objee}{List. EpiEstim object as returned by EpiEstim::estimate_R}

\item{n}{integer. Number of posterior samples drawn.}
}
\description{
Helper function - sample from EpiEstim's R posterior distribution
}
\keyword{internal}
