% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_defaults.R
\name{set_default}
\alias{set_default}
\title{Set a default value in a list}
\usage{
set_default(x, name, value)
}
\arguments{
\item{x}{List.}

\item{name}{String. Name of the list element.}

\item{value}{Default value.}
}
\value{
List.
}
\description{
Set a default value in a list
}
