% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggreg_inc.R
\name{smooth_disaggregation_daily}
\alias{smooth_disaggregation_daily}
\title{Disaggregate reported incidence (typically weekly)
into daily incidence. "Smooth" method.}
\usage{
smooth_disaggregation_daily(cl.data, smooth.input, smooth.input.span)
}
\arguments{
\item{cl.data}{Dataframe of aggregated incidence.
Must have columns named \code{t} and \code{value}.}

\item{smooth.input}{Logical. Should the \emph{input} aggregated incidence
be smoothed before being processed? If TRUE, will perform a
LOESS smoothing.}

\item{smooth.input.span}{Numerical. "span" parameter for the \code{loess()}
smoothing function.}
}
\value{
Dataframe of inferred daily incidence.
}
\description{
Disaggregate reported incidence (typically weekly)
into daily incidence. "Smooth" method.
}
\keyword{internal}
