% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggreg_inc.R
\name{solve_temporal_disagg}
\alias{solve_temporal_disagg}
\title{Title}
\usage{
solve_temporal_disagg(z, obs_times, N, window)
}
\arguments{
\item{z}{Numerical vector. Aggregated (reported) incidence.}

\item{obs_times}{Integer vector. Observation times of the aggregated (reported) incidence.}

\item{N}{Integer. Total number of days after disaggregation.}
}
\value{
Numerical vector of daily disaggregated incidence.
}
\description{
Title
}
\keyword{internal}
