% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_use_dates.R
\name{summarise_by_date}
\alias{summarise_by_date}
\title{Summarise observations by date}
\usage{
summarise_by_date(df)
}
\arguments{
\item{df}{dataframe. has at least \code{date} and \code{mean} columns.}
}
\description{
Summarise observations by date
}
