% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_use_dates.R
\name{summarise_by_date_iters}
\alias{summarise_by_date_iters}
\title{Summarise observations by date for raw iterations from an ensemble}
\usage{
summarise_by_date_iters(df)
}
\arguments{
\item{df}{Data frame. Must have \code{date} and \code{value} columns.}
}
\description{
Summarise observations by date for raw iterations from an ensemble
}
\keyword{internal}
