% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_use_dates.R
\name{summarise_report_counts}
\alias{summarise_report_counts}
\title{Summarise daily inferred reports
based on original reporting schedule and calculate error}
\usage{
summarise_report_counts(df, prm.daily.check)
}
\arguments{
\item{df}{Data frame. As output by \code{\link[=get_use_dates]{get_use_dates()}}.}

\item{prm.daily.check}{List. Parameters for checking aggregated to daily report inference. Elements include:
\itemize{
\item \code{agg.reldiff.tol}: numerical tolerance (\%) for relative error between aggregated inferred daily reports and original aggregated reports; chronological observations are dropped until this tolerance is first acheived (convergence at the start of the timeseries is often the worst, need to maintain uninterrupted daily timeseries for input into Rt calculation).
}
Set this entire argument to \code{NULL} to use inferred daily reports as is.}
}
\description{
Summarise daily inferred reports
based on original reporting schedule and calculate error
}
