% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_postsamples.R
\name{summary_postsamples}
\alias{summary_postsamples}
\title{Helper function that summarises
posterior samples of Rt from EpiEstim}
\usage{
summary_postsamples(x, prm.R)
}
\arguments{
\item{x}{dataframe}

\item{prm.R}{list}
}
\value{
summarised dataframe
}
\description{
Helper function that summarises
posterior samples of Rt from EpiEstim
}
\keyword{internal}
