% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_to_daily.R
\name{weekly_to_daily}
\alias{weekly_to_daily}
\title{Infer daily counts from weekly aggregates}
\usage{
weekly_to_daily(cl.agg, dist.gi, popsize, prm.daily)
}
\arguments{
\item{cl.agg}{dataframe. must have variables \code{date} for the calendar
date of the observation, \code{count} for the count of reported cases.}

\item{dist.gi}{parameters for the generation interval distribution}

\item{popsize}{population size.}

\item{prm.daily}{parameters for daily report inference (via MCMC).}
}
\value{
Dataframe with individual realizations of daily reported cases
}
\description{
Infer daily counts from weekly aggregates
}
