% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww.data}
\alias{ww.data}
\title{Sample of wastewater concentration}
\format{
\subsection{\code{ww.data}}{

A data frame with 47 rows and 3 columns:
\itemize{
\item \code{date}: report date
\item \code{value}: mean sample concentration between multiple replicates
}
}
}
\usage{
ww.data
}
\description{
A subset of SARS-CoV-2 (N2 gene) concentration data in wastewater sampled from the Iona Island wastewater treatment plant in Vancouver between 7 July 2023 and 5 November 2023.
}
\keyword{datasets}
