% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww.input}
\alias{ww.input}
\title{Sample of wastewater concentration}
\format{
\subsection{\code{ww.input}}{

A data frame with 47 rows and 3 columns:
\itemize{
\item \code{date}: report date
\item \code{value}: mean sample concentration between multiple replicates
}
}
}
\usage{
ww.input
}
\description{
A subset of COVID-19 concentration data from the Iona Island wastewater treatment plant in Vancouver.
}
\keyword{datasets}
