% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{agg_to_daily}
\alias{agg_to_daily}
\title{Infer daily counts from aggregates}
\usage{
agg_to_daily(cl.agg, dist.gi, popsize, prm.daily, silent = FALSE)
}
\arguments{
\item{cl.agg}{Data frame. Must have variables:
\itemize{
\item{\code{date}: }{calendar date of report}
\item{\code{count}: }{count of reported cases}
}}

\item{dist.gi}{List. Parameters for generation interval from \code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}.}

\item{popsize}{Integer. Population size to use in MCMC simulation to infer daily observations from aggregated input data.}

\item{prm.daily}{List. Parameters for daily report inference via MCMC. Elements include:
\itemize{
\item{\code{burn}: }{length of burn-in period (number of days)}
\item{\code{iter}: }{number of iterations after burn-in period (number of days)}
\item{\code{chains}: }{number of chains to simulate}
\item{\code{first.agg.period}: }{length of aggregation period for first aggregated observation (number of days); if NULL, assume same aggregation period as observed for second observation (gap between first and second observations)}
}}

\item{silent}{Logical. Should functions be run silently, i.e., without any messages and warnings to console?}
}
\value{
Dataframe with individual realizations of daily reported cases
}
\description{
Infer daily counts from aggregates
}
