% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{agg_to_daily}
\alias{agg_to_daily}
\title{Infer daily counts from aggregates}
\usage{
agg_to_daily(cl.agg, dist.gi, popsize, prm.daily, silent = FALSE)
}
\arguments{
\item{cl.agg}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of report
\item \code{value}: count of reported cases
}}

\item{dist.gi}{List. Parameters for the generation interval distribution in the same format as returned by \code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}.}

\item{popsize}{Integer. Population size to use in MCMC simulation to infer daily observations from aggregated input data.}

\item{prm.daily}{List. Parameters for daily report inference via MCMC. Elements include:
\itemize{
\item \code{burn}: Numeric. Length of burn-in period (number of days).
\item \code{iter}: Numeric. Number of iterations after burn-in period (number of days).
\item \code{chains}: Numeric. Number of chains to simulate.
\item \code{first.agg.period}: length of aggregation period for first aggregated observation (number of days); if NULL, assume same aggregation period as observed for second observation (gap between first and second observations)
}}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\value{
Data frame with individual realizations of daily reported cases
}
\description{
Infer daily counts from aggregates
}
