% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cl-daily.R
\name{attach_t}
\alias{attach_t}
\title{Attach time index column for daily input data}
\usage{
attach_t(cl.input)
}
\arguments{
\item{cl.input}{Data frame. Must have variables:
\itemize{
\item \code{date}: calendar date of report
\item \code{value}: count of reported cases
}}
}
\value{
Tibble.
}
\description{
Attach time index column for daily input data
}
