% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_to_daily.R
\name{attach_t_agg}
\alias{attach_t_agg}
\title{Attach time index (number of days) column
Exclude first day since we don't necessarily know over which period
of time that data was aggregated}
\usage{
attach_t_agg(x)
}
\arguments{
\item{x}{dataframe. must at least have a \code{date} column}
}
\value{
dataframe.
}
\description{
Attach time index (number of days) column
Exclude first day since we don't necessarily know over which period
of time that data was aggregated
}
