% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{attach_t_agg}
\alias{attach_t_agg}
\title{Attach time index (number of days) column
Exclude first day since we don't necessarily know over which period
of time that data was aggregated}
\usage{
attach_t_agg(cl.agg, prm.daily = NULL, silent = FALSE)
}
\arguments{
\item{cl.agg}{Data frame. Must have variables:
\itemize{
\item{\code{date}: }{calendar date of report}
\item{\code{count}: }{count of reported cases}
}}

\item{prm.daily}{List. Parameters for daily report inference via MCMC. Elements include:
\itemize{
\item{\code{burn}: }{length of burn-in period (number of days)}
\item{\code{iter}: }{number of iterations after burn-in period (number of days)}
\item{\code{chains}: }{number of chains to simulate}
\item{\code{first.agg.period}: }{length of aggregation period for first aggregated observation (number of days); if NULL, assume same aggregation period as observed for second observation (gap between first and second observations)}
}}

\item{silent}{Logical. Should functions be run silently, i.e., without any messages and warnings to console?}
}
\value{
dataframe.
}
\description{
Attach time index (number of days) column
Exclude first day since we don't necessarily know over which period
of time that data was aggregated
}
