% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_data_clin}
\alias{check_data_clin}
\title{Check the dataframe of clinical data}
\usage{
check_data_clin(dat, silent = FALSE)
}
\arguments{
\item{dat}{The dataframe of clinical data}

\item{silent}{logical. should the functions be run silently, i.e., without any messages to console?}
}
\description{
Check the dataframe of clinical data
}
