% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.R}
\alias{check_prm.R}
\title{Check parameters for Rt calculation}
\usage{
check_prm.R(x, silent = FALSE)
}
\arguments{
\item{x}{List of parameters for Rt calculation}

\item{silent}{logical. should the functions be run silently, i.e., without any messages to console?}
}
\description{
Check parameters for Rt calculation
}
