% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.R}
\alias{check_prm.R}
\title{Check parameters for Rt calculation}
\usage{
check_prm.R(x, silent = FALSE)
}
\arguments{
\item{x}{List. Parameters for Rt calculation.}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\description{
Check parameters for Rt calculation
}
\keyword{internal}
