% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_prm.R}
\alias{check_prm.R}
\title{Check parameters for Rt calculation}
\usage{
check_prm.R(x)
}
\arguments{
\item{x}{List of parameters for Rt calculation}
}
\description{
Check parameters for Rt calculation
}
