% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{correct_underreporting}
\alias{correct_underreporting}
\title{Correct underreporting by scaling up}
\usage{
correct_underreporting(reports.daily, reporting.fraction)
}
\arguments{
\item{reports.daily}{dataframe of daily reported cases. must at least include \code{value} column with counts.}

\item{reporting.fraction}{numeric. proportion of incidence that is reported.}
}
\description{
Correct underreporting by scaling up
}
