% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{deconv}
\alias{deconv}
\title{Wrapper for deconvolution with a given distribution}
\usage{
deconv(counts, dist, max.iter = 10)
}
\arguments{
\item{dist}{a truncated daily discrete distribution (vector) with which we're deconvoluting the counts}

\item{y}{a vector of daily counts}
}
\description{
Wrapper for deconvolution with a given distribution
}
