% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv_ww_inc.R
\name{deconv_ww_inc}
\alias{deconv_ww_inc}
\title{Deconvoluting Wastewater Data to Incidence}
\usage{
deconv_ww_inc(d, fec, scaling.factor, silent)
}
\arguments{
\item{d}{Data frame. Wastewater dataframe. Must include at least \code{date} and
\code{obs} columns.}

\item{fec}{List. Fecal shedding distribution}

\item{scaling.factor}{Numeric. Scaling from wastewater concentration to
prevalence. This value may be assumed or independently calibrated to data.}

\item{silent}{Logical. Flag to suppress output messages.}
}
\value{
Dataframe with deconvoluted incidence
}
\description{
Function estimates incidence from smoothed data
}
