% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{def_dist_fecal_shedding}
\alias{def_dist_fecal_shedding}
\title{Define fecal Shedding Distribution}
\usage{
def_dist_fecal_shedding(pathogen = "sarscov2", subtype = "")
}
\arguments{
\item{pathogen}{String. Name of the pathogen ('sarscov2', 'influenza')}

\item{subtype}{String.}
}
\value{
A fecal shedding distribution stored as a list.
}
\description{
Define fecal Shedding Distribution
}
\examples{
fec = def_dist_fecal_shedding('sarscov2')
print(fec)

}
