% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{draw_from_gamma}
\alias{draw_from_gamma}
\title{Draw from gamma for a parameter specified in a distribution family list.}
\usage{
draw_from_gamma(par, dist)
}
\arguments{
\item{par}{Character. Name of the parameter to sample.}

\item{dist}{List. Distribution definition, as output by a \verb{def_dist_*()} function.}
}
\value{
Numeric. The sampled parameter value.
}
\description{
Draw from gamma for a parameter specified in a distribution family list.
}
