% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl.R
\name{estimate_R_cl}
\alias{estimate_R_cl}
\title{Estimate the effective reproduction from clinical report data}
\usage{
estimate_R_cl(
  cl.agg,
  dist.repdelay,
  dist.repfrac,
  dist.incub,
  dist.gi,
  popsize,
  prm.daily = list(burn = 500, iter = 2000, chains = 20, first.agg.period = NULL),
  prm.daily.check = list(agg.reldiff.tol = 10),
  prm.smooth = list(window = 7),
  prm.R = list(iter = 10)
)
}
\arguments{
\item{cl.agg}{dataframe. must have variables \code{date} for the calendar
date of the observation, \code{count} for the count of reported cases.}

\item{dist.repdelay}{parameters for the reporting delay distribution}

\item{dist.repfrac}{parameters for the reporting fraction distribution}

\item{dist.incub}{parameters for the incubation period distribution}

\item{dist.gi}{parameters for the generation interval distribution}

\item{popsize}{population size.}

\item{prm.daily}{parameters for daily report inference (via MCMC).}

\item{prm.daily.check}{list. parameters for checking aggregated to daily report inference. set this parameter to \code{NULL} to use inferred daily reports as is. list elements include:
\itemize{
\item \code{agg.reldiff.tol}: numerical tolerance (\%) for relative error between aggregated inferred daily reports and original aggregated reports. observations outside of this tolerance are dropped.
}}

\item{prm.smooth}{smoothing parameters}

\item{prm.R}{settings for the ensemble}
}
\value{
List. Elements include:
\itemize{
\item \code{cl.agg} original aggregated reports signal
\item \code{cl.daily} reports as input for Rt calculation (inferred daily counts, smoothed)
\item \code{R} the effective R estimate (summary from ensemble)
}
}
\description{
Estimate the effective reproduction from clinical report data
}
