% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_rep.R
\name{estimate_R_cl_rep}
\alias{estimate_R_cl_rep}
\title{Ensemble estimate of Rt}
\usage{
estimate_R_cl_rep(
  cl.daily,
  dist.repfrac,
  dist.repdelay,
  dist.incub,
  dist.gi,
  prm.R = list(iter = 100, config.EpiEstim = NULL)
)
}
\arguments{
\item{cl.daily}{dataframe. realizations of daily report counts. must include at least \code{date} and \code{value} columns.}

\item{dist.repfrac}{parameters for the reporting fraction distribution}

\item{dist.repdelay}{parameters for the reporting delay distribution}

\item{dist.incub}{parameters for the incubation period distribution}

\item{dist.gi}{parameters for the generation interval distribution}

\item{prm.R}{list. settings for the ensemble when calculating Rt. elements include:
\itemize{
\item{\code{iter}: }{number of iterations for ensemble}
\item{\code{config.EpiEstim}: }{configuration for \code{EpiEstim} defined via \code{EpiEstim::make_config()}. if \code{NULL}, will use default config from \code{EpiEstim}.}
}}
}
\value{
A summary of individual Rt realizations with 95\% confidence intervals
}
\description{
Ensemble estimate of Rt
}
\seealso{
\code{\link[EpiEstim:make_config]{EpiEstim::make_config()}}
}
