% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_rep.R
\name{estimate_R_cl_rep}
\alias{estimate_R_cl_rep}
\title{Ensemble estimate of Rt}
\usage{
estimate_R_cl_rep(
  cl.input,
  dist.repfrac,
  dist.repdelay,
  dist.incub,
  dist.gi,
  prm.R = list(iter = 100)
)
}
\arguments{
\item{dist.repfrac}{parameters for the reporting fraction distribution}

\item{dist.repdelay}{parameters for the reporting delay distribution}

\item{dist.incub}{parameters for the incubation period distribution}

\item{dist.gi}{parameters for the generation interval distribution}

\item{prm.R}{settings for the ensemble}
}
\value{
A summary of individual Rt realizations with 95\% confidence intervals
}
\description{
Ensemble estimate of Rt
}
