% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{estimate_R_cl_single}
\alias{estimate_R_cl_single}
\title{A single realization of the Rt estimate}
\usage{
estimate_R_cl_single(
  cl.daily,
  dist.repfrac,
  dist.repdelay,
  dist.incub,
  dist.gi,
  prm.R,
  silent = FALSE
)
}
\arguments{
\item{cl.daily}{dataframe. realizations of daily report counts. must include at least \code{date} and \code{value} columns.}

\item{dist.repfrac}{parameters for the reporting fraction distribution}

\item{dist.repdelay}{parameters for the reporting delay distribution}

\item{dist.incub}{parameters for the incubation period distribution}

\item{dist.gi}{parameters for the generation interval distribution}

\item{prm.R}{List. Settings for the ensemble when calculating Rt.
Elements include:
\itemize{
\item \code{window}: number of days defining the window of data used by
\code{EpiEstim} to estimate Rt
\item \code{CI}: Numeric between 0 and 1. Confidence interval width for Rt
estimates after sampling uncertain distributions.
\item{\code{config.EpiEstim}: }{configuration for \code{EpiEstim} defined via
\code{EpiEstim::make_config()}. If \code{NULL}, will use default config from
\code{EpiEstim}.}
}}

\item{silent}{logical. should the functions be run silently, i.e., without any messages to console?}
}
\description{
A single realization of the Rt estimate
}
