% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{estimate_R_cl_single}
\alias{estimate_R_cl_single}
\title{A single realization of the Rt estimate}
\usage{
estimate_R_cl_single(
  cl.daily,
  dist.repfrac,
  dist.repdelay,
  dist.incub,
  dist.gi,
  prm.R
)
}
\arguments{
\item{cl.daily}{dataframe. realizations of daily report counts. must include at least \code{date} and \code{value} columns.}

\item{dist.repfrac}{parameters for the reporting fraction distribution}

\item{dist.repdelay}{parameters for the reporting delay distribution}

\item{dist.incub}{parameters for the incubation period distribution}

\item{dist.gi}{parameters for the generation interval distribution}

\item{prm.R}{list. settings for the ensemble when calculating Rt. elements include:
\itemize{
\item{\code{iter}: }{number of iterations for ensemble}
\item{\code{config.EpiEstim}: }{configuration for \code{EpiEstim} defined via \code{EpiEstim::make_config()}. if \code{NULL}, will use default config from \code{EpiEstim}.}
}}
}
\description{
A single realization of the Rt estimate
}
