% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{estimate_R_cl_single}
\alias{estimate_R_cl_single}
\title{A single realization of the Rt estimate}
\usage{
estimate_R_cl_single(
  cl.daily,
  dist.repfrac,
  dist.repdelay,
  dist.incub,
  dist.gi,
  prm.R,
  silent = FALSE
)
}
\arguments{
\item{dist.repfrac}{List. Parameters for the reporting fraction distribution in the same format as returned by \code{\link[=def_dist_reporting_fraction]{def_dist_reporting_fraction()}}.}

\item{dist.repdelay}{List. Parameters for the reporting delay distribution in the same format as returned by \code{\link[=def_dist_reporting_delay]{def_dist_reporting_delay()}}.}

\item{dist.incub}{List. Parameters for the incubation period distribution in the same format as returned by \code{\link[=def_dist_incubation_period]{def_dist_incubation_period()}}.}

\item{dist.gi}{List. Parameters for the generation interval distribution in the same format as returned by \code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}.}

\item{prm.R}{List. Settings for the ensemble when calculating Rt. Elements include:
\itemize{
\item \code{iter}: number of iterations for ensemble
\item \code{window}: length of time window to use for each Rt estimate. if \code{t_end} is specified in \code{config.EpiEstim}, this option will override it.
\item \code{config.EpiEstim}: configuration for \code{EpiEstim} defined via \code{\link[EpiEstim:make_config]{EpiEstim::make_config()}}. if \code{NULL}, will use default config from \code{EpiEstim}.
}}

\item{silent}{Logical. Flag to suppress output messages.}
}
\description{
A single realization of the Rt estimate
}
