% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_ww.R
\name{estimate_R_ww}
\alias{estimate_R_ww}
\title{Estimate the effective reproduction from wastewater concentration data.}
\usage{
estimate_R_ww(
  ww.conc,
  dist.fec,
  dist.gi,
  scaling.factor = 1,
  prm.smooth = list(window = 14, align = "center", method = "loess", span = 0.2),
  prm.R = list(CI = 0.95, window = 7, config.EpiEstim = NULL),
  iter = 100
)
}
\arguments{
\item{ww.conc}{Dataframe. Must have variables named \code{date} for the wastewater
collection date and \code{val} for the pathogen concentration.}

\item{dist.fec}{Numerical vector. Distribution of fecal shedding (time unit=day).}

\item{dist.gi}{Numerical vector. Distribution of the generation interval (time unit=day).}

\item{scaling.factor}{Numeric. Scaling from wastewater concentration to prevalence.
This value may be assumed or independently calibrated to data.}

\item{prm.smooth}{List. Parameters for the smoothing algorithm of the wastewater signal.}

\item{prm.R}{List. Settings for the ensemble when calculating Rt. Elements include:
\itemize{
\item \code{window}: number of days defining the window of data used by \code{EpiEstim} to estimate Rt
\item \code{CI}: Numeric between 0 and 1. Confidence interval width for Rt estimates after sampling uncertain distributions.
\item{\code{config.EpiEstim}: }{configuration for \code{EpiEstim} defined via \code{EpiEstim::make_config()}. if \code{NULL}, will use default config from \code{EpiEstim}.}
}}

\item{iter}{Integer. Number of samples for the (uncertain) generation interval distribution.}
}
\value{
List. Elements include:
\itemize{
\item \code{ww.conc} original wastewater signal
\item \code{ww.smooth} smoothed wastewater signal
\item \code{inc} inferred incidence
\item \code{R} the effective R estimate
}
}
\description{
Estimate the effective reproduction from wastewater concentration data.
}
