% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_ww.R
\name{estimate_R_ww}
\alias{estimate_R_ww}
\title{Estimate the effective reproduction from wastewater concentration data.}
\usage{
estimate_R_ww(
  ww.conc,
  date.start,
  dist.fec,
  dist.gi,
  scaling.factor = 1,
  prm.smooth = list(window = 14, align = "center", method = "loess", span = 0.2)
)
}
\arguments{
\item{ww.conc}{Dataframe. Must have variables named \code{date} for the wastewater
collection date and \code{value} for the pathogen concentration.}

\item{date.start}{Date. First date of observation used to estimate Rt.}

\item{dist.fec}{Numerical vector. Distribution of fecal shedding (time unit=day).}

\item{dist.gi}{Numerical vector. Distribution of the generation interval (time unit=day).}

\item{scaling.factor}{Numeric. Scaling from wastewater concentration to prevalence.
This value may be assumed or independently calibrated to data.}

\item{prm.smooth}{List. Parameters for the smoothing algorithm of the wastewater signal.}
}
\value{
An object.
}
\description{
Estimate the effective reproduction from wastewater concentration data.
}
\examples{
1+1


}
