% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{fit_jags_aggreg}
\alias{fit_jags_aggreg}
\title{Fit JAGS model to aggregated data}
\usage{
fit_jags_aggreg(
  obs.times,
  Y,
  g,
  N,
  n.days = NULL,
  mcmc.params = list(burn = 1000, iter = 3000, chains = 3),
  silent = FALSE
)
}
\arguments{
\item{obs.times}{numeric. vector of observation times}

\item{Y}{numeric. vector of aggregated counts}

\item{g}{numeric. vector of discretized generation interval density.}

\item{N}{numeric. scalar population size.}

\item{n.days}{numeric. total number of days. if \code{NULL}, use \code{max(obs.times)}}

\item{mcmc.params}{list. MCMC parameters:
\itemize{
\item \code{burn}: burn-in period (days)
\item \code{iter}: iterations after burn-in (days)
\item \code{chains}: number of chains
}}

\item{silent}{Logical. Should functions be run silently, i.e., without any messages and warnings to console?}
}
\description{
Fit JAGS model to aggregated data
}
