% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{fit_jags_aggreg}
\alias{fit_jags_aggreg}
\title{Fit JAGS model to aggregated data}
\usage{
fit_jags_aggreg(obs.times, Y, g, N, n.days = NULL, prm.daily, silent = FALSE)
}
\arguments{
\item{obs.times}{Numeric. Vector of observation times.}

\item{Y}{Numeric. Vector of aggregated counts.}

\item{g}{Numeric. Vector of discretized generation interval density.}

\item{N}{Numeric. Scalar population size.}

\item{n.days}{Numeric. Total number of days. if \code{NULL}, use \code{max(obs.times)}}

\item{prm.daily}{List. Parameters for daily report inference via MCMC. Elements include:
\itemize{
\item \code{burn}: Numeric. Length of burn-in period (number of days). Default is 100.
\item \code{iter}: Numeric. Number of iterations after burn-in period (number of days). Default is 500.
\item \code{chains}: Numeric. Number of chains to simulate. Default is 1.
\item \code{first.agg.period}: length of aggregation period for first aggregated observation (number of days); if NULL, assume same aggregation period as observed for second observation (gap between first and second observations)
}}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\description{
Fit JAGS model to aggregated data
}
