% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{get_discrete_dist}
\alias{get_discrete_dist}
\title{Get a discretized, truncated version of a distribution}
\usage{
get_discrete_dist(params)
}
\arguments{
\item{params}{distribution params (output of \verb{def_dist_*()} function)}
}
\value{
vector with discretized density
}
\description{
Get a discretized, truncated version of a distribution
}
\examples{
prm <- def_dist_incubation_period(); get_discrete_dist(prm)
}
