% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecal_shedding.R
\name{get_gamma_fec}
\alias{get_gamma_fec}
\title{Fecal Shedding Gamma Distribution}
\usage{
get_gamma_fec(fec.df)
}
\arguments{
\item{fec.df}{List of fecal shedding parameters}
}
\value{
A numerical vector where each element represent one day since infection.
}
\description{
Fecal Shedding Gamma Distribution
}
