% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_to_daily.R
\name{get_realizations}
\alias{get_realizations}
\title{Retrieve realizations for weekly -> daily inference}
\usage{
get_realizations(fit.reports.daily, reports)
}
\description{
Retrieve realizations for weekly -> daily inference
}
\seealso{
\code{\link[=weekly_to_daily]{weekly_to_daily()}}
}
