% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_use_dates.R
\name{get_use_dates}
\alias{get_use_dates}
\title{Get dates for which to use (trust) inferred daily reports}
\usage{
get_use_dates(reports.daily, reports)
}
\description{
Get dates for which to use (trust) inferred daily reports
}
