% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{incidence_to_R}
\alias{incidence_to_R}
\title{Estimate Rt using EpiEstim}
\usage{
incidence_to_R(incidence, generation.interval)
}
\arguments{
\item{incidence}{dataframe. estimated incidence. includes at least \code{date}, \code{I}, and \code{t} columns.}

\item{generation.interval}{list. parameters for generation interval from \code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}.}
}
\description{
Estimate Rt using EpiEstim
}
\seealso{
\code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}
}
