% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_to_R.R
\name{incidence_to_R}
\alias{incidence_to_R}
\title{Estimate Rt using EpiEstim}
\usage{
incidence_to_R(incidence, generation.interval, prm.R)
}
\arguments{
\item{incidence}{dataframe. estimated incidence. includes at least \code{date}, \code{I}, and \code{t} columns.}

\item{generation.interval}{list. parameters for generation interval from \code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}.}

\item{prm.R}{list. settings for the ensemble when calculating Rt. elements include:
\itemize{
\item{\code{iter}: }{number of iterations for ensemble}
\item{\code{window}: }{length of time window to use for each Rt estimate. if \code{t_end} is specified in \code{config.EpiEstim}, this option will override it.}
\item{\code{config.EpiEstim}: }{configuration for \code{EpiEstim} defined via \code{EpiEstim::make_config()}. if \code{NULL}, will use default config from \code{EpiEstim}.}
}}
}
\description{
Estimate Rt using EpiEstim
}
\seealso{
\code{\link[=def_dist_generation_interval]{def_dist_generation_interval()}}

\code{\link[EpiEstim:make_config]{EpiEstim::make_config()}}
}
