% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diagnostic_cl}
\alias{plot_diagnostic_cl}
\title{Diagnostic plot for R estimation from clinical data}
\usage{
plot_diagnostic_cl(r.estim)
}
\arguments{
\item{r.estim}{Object as returned from \code{estimate_R_cl()}.}
}
\value{
A ggplot object.
}
\description{
Diagnostic plot for R estimation from clinical data
}
\seealso{
\code{\link[=estimate_R_cl]{estimate_R_cl()}}
}
