% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_diagnostic_ww}
\alias{plot_diagnostic_ww}
\title{Diagnostic plot for R estimation from wastewater}
\usage{
plot_diagnostic_ww(r.estim, caption = NULL)
}
\arguments{
\item{r.estim}{Object as returned from \code{estimate_R_ww()}.}

\item{caption}{String. Optional caption.}
}
\value{
A ggplot object.
}
\description{
Diagnostic plot for R estimation from wastewater
}
\examples{
1+1
}
\seealso{
\code{\link[=estimate_R_ww]{estimate_R_ww()}}
}
