% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot a distribution}
\usage{
plot_dist(d)
}
\arguments{
\item{d}{List that defines the distribution (as returned by \code{def_dist_incubation_period()} for example)}
}
\value{
A ggplot object.
}
\description{
Plot a distribution
}
\examples{
# Define a `ern` distribution:
gi  = def_dist_generation_interval(pathogen = 'influenza')

# Plot can be customized like any `ggplot` object:
g = plot_dist(gi) + ggplot2::labs(subtitle = 'your subtitle')
plot(g)


}
