% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{reports_to_incidence}
\alias{reports_to_incidence}
\title{Infer incidence from reports via a series of deconvolutions}
\usage{
reports_to_incidence(
  reports.daily,
  reporting.delay,
  incubation.period,
  max.iter = 9,
  silent = FALSE
)
}
\arguments{
\item{reports.daily}{Data frame. Daily report counts. Includes at least \code{date} and \code{value} columns.}

\item{reporting.delay}{List. Parameters for a single reporting delay distribution, as generated by \code{\link[=sample_a_dist]{sample_a_dist()}}.}

\item{incubation.period}{List. Parameters for a single incubation period distribution, as generated by \code{\link[=sample_a_dist]{sample_a_dist()}}.}

\item{max.iter}{Numeric. Maximum number of Richardson-Lucy iterations.}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\description{
Infer incidence from reports via a series of deconvolutions
}
\keyword{internal}
