% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R_cl_single.R
\name{reports_to_incidence}
\alias{reports_to_incidence}
\title{Infer incidence from reports via deconvolutions}
\usage{
reports_to_incidence(
  reports.daily,
  reporting.delay,
  incubation.period,
  max.iter = 50,
  silent = FALSE
)
}
\arguments{
\item{reports.daily}{dataframe. daily report counts. includes at least \code{date} and \code{value} columns.}

\item{reporting.delay}{list. parameters for reporting delay distribution (following same format as other \verb{def_dist_*()} function.}

\item{incubation.period}{list. parameters for incubation period from \code{\link[=def_dist_incubation_period]{def_dist_incubation_period()}}.}

\item{max.iter}{numeric. maximum number of Richardson-Lucy iterations}

\item{silent}{logical. should the functions be run silently, i.e., without any messages to console?}
}
\description{
Infer incidence from reports via deconvolutions
}
\seealso{
\code{\link[=def_dist_incubation_period]{def_dist_incubation_period()}}
}
