% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cl.R
\name{smooth_cl}
\alias{smooth_cl}
\title{Smooth realizations from estimating daily reports}
\usage{
smooth_cl(cl.daily, prm.smooth)
}
\arguments{
\item{cl.daily}{Data frame. Output of \code{\link[=agg_to_daily]{agg_to_daily()}}.}

\item{prm.smooth}{List. Smoothing parameters for input signal into Rt calculation. Elements include:
\itemize{
\item{\code{window}: }{width of smoothing window (number of days)}
}}
}
\value{
Dataframe
}
\description{
Smooth realizations from estimating daily reports
}
