% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_cl.R
\name{smooth_cl}
\alias{smooth_cl}
\title{Smooth realizations from estimating daily reports}
\usage{
smooth_cl(cl.daily, prm.smooth)
}
\arguments{
\item{cl.daily}{Dataframe resulting from \code{\link[=weekly_to_daily]{weekly_to_daily()}}}

\item{prm.smooth}{smoothing parameters}
}
\value{
Dataframe
}
\description{
Smooth realizations from estimating daily reports
}
