% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ww.R
\name{smooth_ww}
\alias{smooth_ww}
\title{Smooth wastewater data}
\usage{
smooth_ww(ww.conc, prm.smooth, silent = FALSE)
}
\arguments{
\item{ww.conc}{Data frame. Must have variables named \code{date} for the
wastewater collection date and \code{value} for the pathogen concentration.}

\item{prm.smooth}{List. list of smoothing parameters. Parameters should be
specified as followed:
\enumerate{
\item method - smoothing method (either 'rollmean' or 'loess')
\item window - smoothing window (number of days for rollmean smoothing)
\item align - smoothing alignment (for rollmean smoothing, either
'center', 'left', 'right')
\item span - smoothing span (for loess smoothing)
}
Set this entire list to \code{NULL} to turn off smoothing}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\value{
Data frame
}
\description{
Function takes pre-processed wastewater data and smooths it
}
