% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ww.R
\name{smooth_ww}
\alias{smooth_ww}
\title{Smoothing ww data}
\usage{
smooth_ww(df, date.start, prm.smooth)
}
\arguments{
\item{df}{ww dataframe}

\item{prm.smooth}{list of smoothing parameters}

\item{start.date}{desired start date}
}
\value{
dataframe with smoothed ww data
}
\description{
Function takes cleaned ww data and smooths data
}
