% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ww.R
\name{smooth_ww}
\alias{smooth_ww}
\title{Smooth wastewater data}
\usage{
smooth_ww(ww.conc, prm.smooth, silent = FALSE)
}
\arguments{
\item{ww.conc}{Data frame. Must have variables named \code{date} for the
wastewater collection date and \code{value} for the pathogen concentration.}

\item{prm.smooth}{List. list of smoothing parameters. Parameters should be
specified as followed:
\itemize{
\item \code{method}: smoothing method, either \code{'rollmean'} (rolling mean) or \code{'loess'} (LOESS smoothing)
\item \code{window}: for \verb{method = 'rollmean} only; width of smoothing window in days
\item \code{align}: for \verb{method = 'rollmean} only; smoothing alignment, either \code{'center'}, \code{'left'}, \code{'right'}
\item \code{span}: for \code{method = 'loess'} only; smoothing span (see the documentation for \code{stats::loess()} for details)
}
Set this entire list to \code{NULL} to turn off smoothing}

\item{silent}{Logical. Flag to suppress all output messages, warnings, and progress bars.}
}
\value{
Data frame
}
\description{
Function takes pre-processed wastewater data and smooths it
}
